+++ src/briefy/leica/tools/finance_csv_export.py
@@ -88,6 +88,8 @@ def export_order(state=None, customer_comments=False):
         'price_currency',
         'price',
         'delivery_sftp_link',
+        'type',
+        'actual_order_price',
     ]
 
     query = Order.query()
@@ -146,7 +148,9 @@ def export_order(state=None, customer_comments=False):
             accept_date=export_datetime(accept_date),
             price_currency=item.price_currency,
             price=export_integer(item.price),
-            delivery_sftp_link=delivery_sftp_link
+            delivery_sftp_link=delivery_sftp_link,
+            type=item.type,
+            actual_order_price=export_integer(item.actual_order_price)
         )
 
         if customer_comments: